pyproject.toml
[project]
name = "tg2pdf"
version = "0.1.0"
description = "Converts Telegram posts to pdf"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "telethon>=1.40.0",
]
